# internal/gorm/dbal/models/project_li.go
package models

// Code generated by a tool. DO NOT EDIT.
// https://sqldalmaker.sourceforge.net/

// ProjectLi `project` list item: [+] p_tasks_count
type ProjectLi struct {
	PID         int64  `json:"p_id" gorm:"column:p_id;primaryKey;autoIncrement;not null"`
	PName       string `json:"p_name" gorm:"column:p_name;size:256;not null"`
	PTasksCount int64  `json:"p_tasks_count" gorm:"column:p_tasks_count;not null"`
}