# internal/gorm/dbal/models/task_li.go
package models

// Code generated by a tool. DO NOT EDIT.
// https://sqldalmaker.sourceforge.net/

// TaskLi `task` list item: [-] t_comments
type TaskLi struct {
	TID       int64  `json:"t_id" gorm:"column:t_id;primaryKey;autoIncrement;not null"`
	PID       int64  `json:"p_id" gorm:"column:p_id"`
	TPriority int64  `json:"t_priority" gorm:"column:t_priority;not null"`
	TDate     string `json:"t_date" gorm:"column:t_date;not null"`
	TSubject  string `json:"t_subject" gorm:"column:t_subject;not null"`
}