# internal/sqlx/dbal/dto/project.go
package dto

// Code generated by a tool. DO NOT EDIT.
// https://sqldalmaker.sourceforge.net/

type Project struct {
	PID   int64  `json:"p_id" db:"p_id"`
	PName string `json:"p_name" db:"p_name"`
}