# internal/sqlx/dbal/dto/project_li.go
package dto

// Code generated by a tool. DO NOT EDIT.
// https://sqldalmaker.sourceforge.net/

// ProjectLi `project` list item: [+] p_tasks_count
type ProjectLi struct {
	PID         int64  `json:"p_id" db:"p_id"`
	PName       string `json:"p_name" db:"p_name"`
	PTasksCount int64  `json:"p_tasks_count" db:"p_tasks_count"`
}